refactor: Extract getPosts method
