+++ src/lib/utils/index.ts
@@ -1,6 +1,13 @@
 import closeMenuIfOpen from './closeMenuIfOpen';
 import formatDate from './formatDate';
+import getPosts from './getPosts';
 import removeScrollLock from './removeScrollLock';
 import toggleScrollLock from './toggleScrollLock';
 
-export { closeMenuIfOpen, formatDate, removeScrollLock, toggleScrollLock };
+export {
+	closeMenuIfOpen,
+	formatDate,
+	getPosts,
+	removeScrollLock,
+	toggleScrollLock
+};
